+++ backend/src/Application.php
@@ -27,13 +27,16 @@ use Neucore\Command\UpdateCorporations;
 use Neucore\Command\UpdateMemberTracking;
 use Neucore\Command\UpdatePlayerGroups;
 use Neucore\Exception\RuntimeException;
+use Neucore\Factory\RepositoryFactory;
 use Neucore\Log\Context;
 use Neucore\Middleware\Psr15\Cors;
 use Neucore\Middleware\Psr15\BodyParams;
+use Neucore\Middleware\Psr15\RateLimit;
 use Neucore\Slim\SessionMiddleware;
 use Neucore\Service\AppAuth;
 use Neucore\Service\Config;
 use Neucore\Service\UserAuth;
+use Neucore\Storage\StorageInterface;
 use Psr\Container\ContainerExceptionInterface;
 use Psr\Container\ContainerInterface;
 use Psr\Http\Message\ResponseFactoryInterface;
@@ -264,6 +267,14 @@ class Application
     private function addMiddleware(App $app): void
     {
         // Add middleware, last added are executed first.
+
+        $app->add(new RateLimit(
+            $this->container->get(AppAuth::class),
+            $this->container->get(StorageInterface::class),
+            $this->container->get(ResponseFactoryInterface::class),
+            $this->container->get(LoggerInterface::class),
+            $this->container->get(RepositoryFactory::class)
+        ));
 
         /** @noinspection PhpIncludeInspection */
         $app->add(new SecureRouteMiddleware(