+++ frontend/src/pages/SystemSettings--Features.vue
@@ -1,7 +1,7 @@
 <template>
 <div class="card border-secondary mb-3">
+    <div class="card-header">Groups Deactivation</div>
     <div class="card-body">
-        <strong>Deactivate Groups</strong>
         <div class="custom-control custom-checkbox">
             <input class="custom-control-input" type="checkbox" value="1"
                    id="groups_require_valid_token" name="groups_require_valid_token"
@@ -14,8 +14,8 @@
                 are excluded from this.
             </label>
         </div>
-        <label class="mt-2">
-            <input type="text" pattern="[0-9]*" class="form-control input-delay"
+        <label class="mt-2 display-block">
+            <input type="text" pattern="[0-9]*" class="form-control input-inline"  name="account_deactivation_delay"
                    v-model="settings.account_deactivation_delay"
                    v-on:input="$emit('changeSettingDelayed', 'account_deactivation_delay', $event.target.value)">
             Delay the deactivation after a token became invalid (hours).
@@ -38,31 +38,59 @@
                          placeholder="Select corporations">
             </multiselect>
         </div>
+    </div>
 
-        <hr>
-
+    <div class="card-header">"Managed" Logins</div>
+    <div class="card-body">
         <div class="custom-control custom-checkbox">
             <input class="custom-control-input" type="checkbox" value="1"
                    id="allow_login_managed" name="allow_login_managed"
                    :checked="settings.allow_login_managed === '1'"
                    @change="$emit('changeSetting', 'allow_login_managed', $event.target.checked ? '1' : '0')">
             <label class="custom-control-label" for="allow_login_managed">
-                <strong>Allow "managed" Login:</strong>
                 Enables the login URL for managed accounts that do not require ESI scopes.
             </label>
         </div>
+    </div>
 
-        <hr>
-
+    <div class="card-header">Character Deletion</div>
+    <div class="card-body">
         <div class="custom-control custom-checkbox">
             <input class="custom-control-input" type="checkbox" value="1"
                    id="allow_character_deletion" name="allow_character_deletion"
                    :checked="settings.allow_character_deletion === '1'"
                    @change="$emit('changeSetting', 'allow_character_deletion', $event.target.checked ? '1' : '0')">
             <label class="custom-control-label" for="allow_character_deletion">
-                <strong>Delete Characters:</strong>
                 Check to allow users to delete their characters.
             </label>
+        </div>
+    </div>
+
+    <div class="card-header">API Rate Limit for Apps</div>
+    <div class="card-body">
+        <p>
+            If configured, each response will contain the headers 'X-Neucore-Rate-Limit-Remain' and
+            'X-Neucore-Rate-Limit-Reset'. If enabled, each request results in error 429 "Too many requests"
+            if the limit has been exceeded.
+        </p>
+        <label class="mt-2 display-block">
+            <input type="text" pattern="[0-9]*" class="form-control input-inline" name="api_rate_limit_max_requests"
+                   v-model="settings.api_rate_limit_max_requests"
+                   v-on:input="$emit('changeSettingDelayed', 'api_rate_limit_max_requests', $event.target.value)">
+            Maximum requests.
+        </label>
+        <label class="mt-2 display-block">
+            <input type="text" pattern="[0-9]*" class="form-control input-inline" name="api_rate_limit_reset_time"
+                   v-model="settings.api_rate_limit_reset_time"
+                   v-on:input="$emit('changeSettingDelayed', 'api_rate_limit_reset_time', $event.target.value)">
+            Reset time in seconds.
+        </label>
+        <div class="custom-control custom-checkbox">
+            <input class="custom-control-input" type="checkbox" value="1"
+                   id="api_rate_limit_active" name="api_rate_limit_active"
+                   :checked="settings.api_rate_limit_active === '1'"
+                   @change="$emit('changeSetting', 'api_rate_limit_active', $event.target.checked ? '1' : '0')">
+            <label class="custom-control-label" for="api_rate_limit_active">Active.</label>
         </div>
     </div>
 </div>
@@ -146,8 +174,11 @@ function readSettings(vm) {
 </script>
 
 <style type="text/scss" scoped>
-    .input-delay {
+    .input-inline {
         display: inline;
         width: 100px;
+    }
+    .display-block {
+        display: block;
     }
 </style>